requirements.txt
djangorestframework
djangorestframework-simplejwt
django-cors-headers
black
stripe
asgiref
Django
pytz
sqlparse
gunicorn
dj-database-url
psycopg2-binary
whitenoise
python-dotenv
django-rest-passwordreset
pillow
django-cloudinary-storage
django-quill-editor
django-anymail[postmark]
